Remove offers from my loans page Rename useTokenOffersOptimisticBN to useTokenOffersOptimistic Remove old useTokenOffersOptimistic
