+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/ActionsCell/ActionsCell.tsx
@@ -2,7 +2,6 @@ import { FC } from 'react'
 
 import { Button } from '@banx/components/Buttons'
 
-import { Offer } from '@banx/api/nft'
 import { core } from '@banx/api/tokens'
 import { useModal } from '@banx/store/common'
 import { isTokenLoanTerminating } from '@banx/utils'
@@ -14,7 +13,6 @@ import styles from './ActionsCell.module.less'
 
 interface ActionsCellProps {
   loan: core.TokenLoan
-  offers: Record<string, Offer[]>
   isCardView: boolean
   disableActions: boolean
 }
@@ -23,12 +21,6 @@ export const ActionsCell: FC<ActionsCellProps> = ({ loan, isCardView, disableAct
   const { open } = useModal()
 
   const isLoanTerminating = isTokenLoanTerminating(loan)
-
-  // const refinanceAvailable = useMemo(() => {
-  //   const offersByMarket = offers[loan.fraktBond.hadoMarket || '']
-  //   return !!offersByMarket?.filter(isOfferNotEmpty).length
-  // }, [offers, loan])
-
   const buttonSize = isCardView ? 'default' : 'small'
 
   return (
